Added in redux-persist, enabling it to work with immutable stores.
As a requirement for being able to persist environments across sessions
we need a module that can easily be dropped in to work for our requirements.

Still more work required for whitelisting variables within a reducer as
well as adding a <PersistGate /> like component to hold until the store
has been re-hydrated.